Adicionar parâmetro para indicar o diretório de instalação

+++ aseprite-uninstaller.sh
@@ -26,12 +26,38 @@ echo -e "\e[1;33m
 ----------------------------------------------------
 "
 
-sudo rm -r ~/aseprite
-sudo rm -r ~/deps
-sudo rm $HOME/.local/share/applications/aseprite.desktop
-sudo rm $HOME/.local/share/icons/hicolor/16x16/apps/aseprite.png
-sudo rm $HOME/.local/share/icons/hicolor/32x32/apps/aseprite.png
-sudo rm $HOME/.local/share/icons/hicolor/48x48/apps/aseprite.png
-sudo rm $HOME/.local/share/icons/hicolor/64x64/apps/aseprite.png
-sudo rm $HOME/.local/share/icons/hicolor/128x128/apps/aseprite.png
-sudo rm $HOME/.local/share/icons/hicolor/256x256/apps/aseprite.png
+DIR_INSTALACAO=$HOME/aseprite
+ARGS=$(getopt -a -n aseprite-installer.sh -o f: --long dir-instalacao: -- "$@")
+eval set -- "$ARGS"
+while :
+do
+  case "$1" in
+    --dir-instalacao) 
+        DIR_INSTALACAO="$2"
+        shift 2 ;;
+    --) shift; break ;;
+  esac
+done
+
+if ! [ -d $DIR_INSTALACAO ]
+then
+  echo "$DIR_INSTALACAO não existe. Abortando..."
+  exit 1
+fi
+
+read -p "Você têm certeza que deseja desinstalar o aseprite localizado no diretório $DIR_INSTALACAO? (Y/N): " -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+  sudo rm -r ~/aseprite
+  sudo rm -r ~/deps
+  sudo rm $HOME/.local/share/applications/aseprite.desktop
+  sudo rm $HOME/.local/share/icons/hicolor/16x16/apps/aseprite.png
+  sudo rm $HOME/.local/share/icons/hicolor/32x32/apps/aseprite.png
+  sudo rm $HOME/.local/share/icons/hicolor/48x48/apps/aseprite.png
+  sudo rm $HOME/.local/share/icons/hicolor/64x64/apps/aseprite.png
+  sudo rm $HOME/.local/share/icons/hicolor/128x128/apps/aseprite.png
+  sudo rm $HOME/.local/share/icons/hicolor/256x256/apps/aseprite.png
+else
+  echo "Abortando..."
+fi
